Send username and password

+++ models/admin/Teams.php
@@ -36,16 +36,16 @@ class Teams extends BLL{
            return parent::insert();
 
         }
-        
+
         function getNewProfileSerial(){
             $obj=new \App\Models\Profile\Profile;
             //echo $this->serial."-";
             $last=$obj->supperUser()->where('serial','llike',$this->serial)->withDeleted()->orderBy("serial","desc")->limit(1)->get();
-             
+
             if(count($last)>0){
                    // $fmt = new \NumberFormatter( 'de_DE', \NumberFormatter::TYPE_INT64);
-                    $newsn=intval($last[0]->serial);//$fmt->parse($last[0]->serial, \NumberFormatter::TYPE_INT64);
-                    
+                $newsn=floatval($last[0]->serial);//$fmt->parse($last[0]->serial, \NumberFormatter::TYPE_INT64);
+
                    // echo $last[0]->serial,"-",$newsn;
                    // exit();
  			        return str_pad($newsn+1, strlen($last[0]->serial), "0", STR_PAD_LEFT);
@@ -53,7 +53,7 @@ class Teams extends BLL{
                 return $this->serial.str_pad(1, 2, "0", STR_PAD_LEFT);
             }
         }
-        
+
 		function getNewSerial($parentid){
 			$obj=new \App\Models\Admin\Teams;
             if($parentid->id>0){
@@ -80,7 +80,7 @@ class Teams extends BLL{
                     //$fmt = new \NumberFormatter( 'de_DE', \NumberFormatter::DECIMAL );
                     //$newsn=$fmt->parse($last[0]->serial, \NumberFormatter::TYPE_INT32);
                     $newsn=intval($last[0]->serial);
-                    
+
 			        return str_pad($newsn+1, $length, "0", STR_PAD_LEFT);
             }else{
                 return $parentid->serial.str_pad(1, $level_array[$parentid->level_type+1], "0", STR_PAD_LEFT);
